feat: implemented a cache for the generated token, allowing for us to hold it for 1 hour which is how long it lasts, therefore making it so that we regenerate it after the hour is up, which is when the token expires, chore: moved some stuff around so that it is better organized
