+++ back-end/src/test/java/edu/brown/cs/student/Tests/server/spotify/handlerTests/recommendationHandlerTests/MockedRecommendationHandlerTests.java
@@ -1,4 +1,4 @@
-package edu.brown.cs.student.Tests.server.spotify.handlerTests;
+package edu.brown.cs.student.Tests.server.spotify.handlerTests.recommendationHandlerTests;
 
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
